Improve icp configuration instructions
